+++ projectile-export/src/main/scala/com/kyleu/projectile/models/feature/graphql/thrift/ThriftSchemaInputHelper.scala
@@ -18,7 +18,7 @@ object ThriftSchemaInputHelper {
     case IntegerType => "IntType"
     case LongType => "LongType"
     case FloatType => "FloatType"
-    case DoubleType => "DoubleType"
+    case DoubleType => "FloatType"
     case BigDecimalType => "BigDecimalType"
 
     case DateType => "localDateType"
